fixed bug focus comment not stripped
